+++ spec/fixtures/cookbooks/test_gitlab_pgbouncer_user/metadata.rb
@@ -1,0 +1,2 @@
+name 'test_gitlab_pgbouncer_user'
+depends 'gitlab-ee'